Moved exportData.js to new services directory.
